services/analytics/data_processor/requirements.txt
redis==5.0.5
ta-lib==0.6.4
pandas_ta==0.4.71b0
clickhouse-connect==0.8.15
pyarrow==19.0.1
click==8.1.8
